Moving Row files out of their folders (since field array has already been removed)
