+++ frontend/src/components/business/datarow/ContainersSourcingRow.js
@@ -1,6 +1,6 @@
 import IconButton from "@mui/material/IconButton";
 import RemoveIcon from "@mui/icons-material/RemoveCircleOutline";
-import TextField from "../../../common/form/TextField";
+import TextField from "../../common/form/TextField";
 import { useFormContext } from "react-hook-form";
 
 const ContainersSourcingRow = ({ index = -1, removeFn = () => {} }) => {
